add commands whitelist and delete

+++ events/messageCreate.js
@@ -1,7 +1,7 @@
 const { Collection } = require('discord.js');
+const { whiteList } = require('../config.json');
 
 const links = new Collection();
-console.log(links);
 
 module.exports = {
   name: 'messageCreate',
@@ -9,7 +9,8 @@ module.exports = {
     // Check if message is a link
     if (message.content.match(/\b(http(s?)):\/\/\S+/gi)) {
       const link = message.content.match(/\b(http(s?)):\/\/\S+/gi)[0];
-
+      // Return if message includes a whitelisted link
+      if (whiteList.some((item) => link.includes(item))) return;
       // Check if link is in collection
       if (links.has(link)) {
         message.delete();
